fix: input focus, send button states, header toolbar, welcome message
- Remove ugly focus border on chat textarea
- Send button always visible: disabled→enabled→stop based on state
- Arrow icon left-to-right (not up)
- Video + ellipsis buttons moved to shared header row
- Chat welcome message uses widget_body not widget_greeting
